Refactor d2l-switch to enable d2l-switch-visibility.
