+++ components/switch/demo/switch.html
@@ -10,6 +10,7 @@
 			import '../../demo/demo-page.js';
 			import '../../tooltip/tooltip.js';
 			import '../switch.js';
+			import '../switch-visibility.js';
 		</script>
 	</head>
 	<body unresolved>
@@ -36,11 +37,11 @@
 
 			<d2l-demo-snippet>
 				<template>
-					<d2l-switch label="Visibility" on-icon="tier1:visibility-show" off-icon="tier1:visibility-hide"></d2l-switch>
+					<d2l-switch-visibility></d2l-switch-visibility>
 				</template>
 			</d2l-demo-snippet>
 
-			<h2>Visibility (with tooltip)</h2>
+			<h2>On/Off (with tooltip)</h2>
 
 			<d2l-demo-snippet>
 				<template>
