+++ components/switch/switch-mixin.js
@@ -1,0 +1,141 @@
+import '../colors/colors.js';
+import { css, html } from 'lit-element/lit-element.js';
+
+export const SwitchMixin = superclass => class extends superclass {
+
+	static get properties() {
+		return {
+			label: { type: String },
+			labelHidden: { type: Boolean, attribute: 'label-hidden' },
+			on: { type: Boolean, reflect: true },
+			role: { type: String, reflect: true },
+			tabindex: { type: String, reflect: true }
+		};
+	}
+
+	static get styles() {
+		return css`
+			:host {
+				display: inline-block;
+				outline-style: none;
+			}
+			:host([hidden]) {
+				display: none;
+			}
+			.d2l-switch-container {
+				background-color: #ffffff;
+				border: 2px solid transparent;
+				border-radius: 1rem;
+				box-sizing: border-box;
+				display: inline-block;
+				font-size: 0;
+				line-height: 0;
+				padding: 0.1rem;
+				vertical-align: middle;
+			}
+			.d2l-switch-inner {
+				border: 1px solid var(--d2l-color-ferrite);
+				border-radius: 0.8rem;
+				box-sizing: border-box;
+				padding: 0.3rem;
+			}
+			:host([on]) .d2l-switch-inner {
+				background-color: var(--d2l-color-celestine-plus-2);
+				border-color: var(--d2l-color-celestine);
+			}
+			.d2l-switch-toggle {
+				position: relative;
+				transition: transform 150ms linear;
+			}
+			.d2l-switch-toggle > div {
+				background-color: #ffffff;
+				border: 1px solid var(--d2l-color-ferrite);
+				border-radius: 0.6rem;
+				box-sizing: border-box;
+				display: inline-block;
+				height: 1.2rem;
+				left: -0.2rem;
+				position: absolute;
+				top: -0.2rem;
+				width: 1.2rem;
+			}
+			:host([on]) .d2l-switch-toggle {
+				transform: translateX(1.2rem);
+			}
+			:host([on]) .d2l-switch-toggle > div {
+				border-color: var(--d2l-color-celestine);
+			}
+			:host(:focus) .d2l-switch-container,
+			:host(:hover) .d2l-switch-container {
+				border-color: var(--d2l-color-celestine);
+				border-width: 2px;
+			}
+			d2l-icon, d2l-icon-custom {
+				height: 0.8rem;
+				width: 0.8rem;
+			}
+			.d2l-switch-icon-on, .d2l-switch-icon-off {
+				display: inline-block;
+			}
+			.d2l-switch-icon-on {
+				margin-right: 0.4rem;
+			}
+			:host([on]) .d2l-switch-icon-on > d2l-icon,
+			:host([on]) .d2l-switch-icon-on > d2l-icon-custom {
+				color: var(--d2l-color-celestine);
+			}
+			@media (prefers-reduced-motion: reduce) {
+				.d2l-switch-toggle {
+					transition: none;
+				}
+			}
+		`;
+	}
+
+	constructor() {
+		super();
+		this.labelHidden = false;
+		this.on = false;
+		this.role = 'switch';
+		this.tabindex = '0';
+	}
+
+	firstUpdated(changedProperties) {
+		super.firstUpdated(changedProperties);
+		this.addEventListener('click', () => {
+			this.on = !this.on;
+		});
+		this.addEventListener('keydown', (e) => {
+			// space pressed... prevent defaul browser scroll
+			if (e.keyCode === 32) e.preventDefault();
+		});
+		this.addEventListener('keyup', (e) => {
+			if (e.keyCode === 32) this.on = !this.on;
+		});
+	}
+
+	render() {
+		return html`
+			<div class="d2l-switch-container">
+				<div class="d2l-switch-inner">
+					<div class="d2l-switch-toggle"><div></div></div>
+					<div class="d2l-switch-icon-on">${this.onIcon}</div>
+					<div class="d2l-switch-icon-off">${this.offIcon}</div>
+				</div>
+			</div>
+			${!this.labelHidden ? html`<span aria-hidden="true">${this.label}</span>` : ''}
+		`;
+	}
+
+	updated(changedProperties) {
+		super.updated(changedProperties);
+		if (changedProperties.has('label')) {
+			this.setAttribute('aria-label', this.label);
+		}
+		if (changedProperties.has('on')) {
+			if (this.on) this.setAttribute('aria-checked', 'true');
+			else this.setAttribute('aria-checked', 'false');
+		}
+	}
+
+};
